requirements.txt
tensorflow-cpu==2.17.0
streamlit
Pillow
numpy
